Moved reusable functionality into separate file
